# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.BuyerAttributes:
    properties:
      card_number_id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
    type: object
  domain.BuyerDoc:
    properties:
      card_number_id:
        type: integer
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
  domain.BuyerListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.BuyerDoc'
        type: array
    type: object
  domain.BuyerPurchaseOrdersReport:
    properties:
      buyer_id:
        type: integer
      card_number_id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      purchase_orders_count:
        type: integer
    type: object
  domain.BuyerPurchaseOrdersReportResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.BuyerPurchaseOrdersReport'
        type: array
    type: object
  domain.BuyerResponse:
    properties:
      data:
        $ref: '#/definitions/domain.BuyerDoc'
    type: object
...
  domain.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestionar sellers y warehouses
  title: API de Sellers y Warehouses
  version: "1.0"
paths:
  /buyers:
    get:
      description: Devuelve todos los buyers del sistema
      produces:
      - application/json
      responses:
        "200":
          description: Lista de buyers obtenida exitosamente
          schema:
            $ref: '#/definitions/domain.BuyerListResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Devuelve todos los buyers
      tags:
      - Buyer
    post:
      consumes:
      - application/json
      description: Crea un nuevo buyer en el sistema
      parameters:
      - description: Buyer a crear
        in: body
        name: buyer
        required: true
        schema:
          $ref: '#/definitions/domain.BuyerAttributes'
      produces:
      - application/json
      responses:
        "201":
          description: Buyer creado exitosamente
          schema:
            $ref: '#/definitions/domain.BuyerDoc'
        "409":
          description: Código del buyer ya existente
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "422":
          description: Error de validación
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Crea un nuevo buyer
      tags:
      - Buyer
  /buyers/{id}:
    delete:
      description: Elimina un buyer por id
      parameters:
      - description: ID del buyer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Buyer eliminado exitosamente
        "400":
          description: Error de validación
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Buyer no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Elimina un buyer por id
      tags:
      - Buyer
    get:
      description: Devuelve un buyer por id
      parameters:
      - description: ID del buyer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Buyer obtenido exitosamente
          schema:
            $ref: '#/definitions/domain.BuyerResponse'
        "400":
          description: Error de validación
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Buyer no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Devuelve un buyer por id
      tags:
      - Buyer
    patch:
      consumes:
      - application/json
      description: Actualiza un buyer por id
      parameters:
      - description: ID del buyer
        in: path
        name: id
        required: true
        type: integer
      - description: Buyer a actualizar
        in: body
        name: buyer
        required: true
        schema:
          $ref: '#/definitions/domain.BuyerAttributes'
      produces:
      - application/json
      responses:
        "200":
          description: Buyer actualizado exitosamente
          schema:
            $ref: '#/definitions/domain.BuyerResponse'
        "400":
          description: Error de validación
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Buyer no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Código del buyer ya existente
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Actualiza un buyer por id
      tags:
      - Buyer
  /buyers/reportPurchaseOrders:
    get:
      description: Devuelve un reporte de órdenes de compra por Buyer
      parameters:
      - description: ID del buyer
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reporte de órdenes de compra obtenido exitosamente
          schema:
            $ref: '#/definitions/domain.BuyerPurchaseOrdersReportResponse'
        "400":
          description: Error de validación
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Buyer no encontrado
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Reporte de órdenes de compra por Buyer
      tags:
      - Buyer
...
swagger: "2.0"